Misc updates

+++ prepare.sh
@@ -14,9 +14,12 @@ uv run otc2otf $IOS_FONT_NAME.ttc
 mv AppleColorEmoji.ttf $WORK_DIR/${IOS_FONT_NAME}_00.ttf
 mv .AppleColorEmojiUI.ttf $WORK_DIR/${IOS_FONT_NAME}_01.ttf
 
-uv run otc2otf $MAC_FONT_NAME.ttc
-mv AppleColorEmoji.ttf $WORK_DIR/${MAC_FONT_NAME}_00.ttf
-rm .AppleColorEmojiUI.ttf
+if [ -f $MAC_FONT_NAME.ttc ]
+then
+    uv run otc2otf $MAC_FONT_NAME.ttc
+    mv AppleColorEmoji.ttf $WORK_DIR/${MAC_FONT_NAME}_00.ttf
+    rm .AppleColorEmojiUI.ttf
+fi
 
 cd $WORK_DIR
 echo "Extracting tables..."
@@ -25,7 +28,11 @@ UNSHARED_TABLES='-t head -t hhea -t meta -t name -t trak -t cntr'
 SHARED_TABLES='-x GDEF -x DSIG -x cmap -x feat -x glyf -x hmtx -x loca -x maxp -x morx -x post -x sbix -x vhea -x vmtx -x GPOS -x GlyphOrder -x OS/2'
 uv run ttx -q -s -f -x DSIG ${IOS_FONT_NAME}_00.ttf
 uv run ttx -q -s -f $SHARED_TABLES ${IOS_FONT_NAME}_01.ttf
-uv run ttx -q -s -f -t sbix ${MAC_FONT_NAME}_00.ttf
+
+if [ -f $MAC_FONT_NAME.ttc ]
+then
+    uv run ttx -q -s -f -t sbix ${MAC_FONT_NAME}_00.ttf
+fi
 
 echo "Fixing up interracial emojis..."
 uv run emojifonts-shift-multi ${IOS_FONT_NAME}_00._h_m_t_x.ttx
